ceph-salt-formula: eliminate implicit dependency on which
"which" is not a shell builtin; it is distributed as a separate RPM.

The same functionality (testing whether a given executable is available
in the system) can be obtained using the "type" shell builtin.

Please note that this commit is not introducing a bashism:
"sh -c 'type foo'" should work with any shell.

Signed-off-by: Nathan Cutler <ncutler@suse.com>