Optionally deploy monitoring stack on bootstrap node
d59c8eb9d1e96aa9d7424e0a236471019d84ae6e added "--skip-monitoring-stack"
to the "cephadm bootstrap" command, which had the effect of
unconditionally removing the user's access to this cephadm feature.

There might be scenarios where the user would want to deploy the
monitoring stack, so add a configurable for it.

Fixes: https://github.com/ceph/ceph-salt/issues/274
Signed-off-by: Nathan Cutler <ncutler@suse.com>