cephbootstrap: run bootstrap with --apply-spec
As reported in the downstream bug cited below, "cephadm bootstrap"
is now deploying 5 MONs and 2 MGRs by default.

The expectation is that "ceph-salt apply" will cause one (and only one)
MON and one (and only one) MGR to be deployed, and that both the MON and
the MGR will be deployed on the node specified by the user as the
bootstrap node.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1174422
Signed-off-by: Nathan Cutler <ncutler@suse.com>