execute: check for running jobs when failing to re-apply formula
Originally, we thought that the only case when

    salt -G ceph-salt:member state.sls_exists ceph-salt

can fail is when the ceph-salt-formula package is not installed. Now we
know it can also fail when jobs initiated by an earlier invocation of
the Salt Formula are still running.

Check for such jobs and display a different error message in this case.

Fixes: https://github.com/ceph/ceph-salt/issues/327
Signed-off-by: Nathan Cutler <ncutler@suse.com>