execute: implement test.ping sanity check
Before this commit, if any minion is down or otherwise not responding to
the Salt Master, "ceph-salt apply" will fail with a bogus error message:

"Could not find ceph-salt formula. Please check if ceph-salt-formula
package is installed"

To address this, all configured minions are pinged and number of
responding minions is compared against the number of configured minions.
If any do not respond, an appropriate error message is displayed.

Fixes: https://github.com/ceph/ceph-salt/issues/331
Signed-off-by: Nathan Cutler <ncutler@suse.com>