Sanity-check time sync services when /time_server is disabled
Users can set /time_server to "disabled", but if they do that,
they are responsible for ensuring that the time sync service is enabled
and running on all the minions before starting to deploy a Ceph cluster.

Now, you may ask: "cephadm check-host" already checks this, so why are
we duplicating that check here in ceph-salt?

The answer is: without a pre-flight check for this, a user might disable
/time_server and then run "ceph-salt apply", and in this case "ceph-salt
apply" will fail while trying to apply the Salt Formula, yet we are
trying to catch all possible failure conditions before actually applying
the Salt Formula...

Fixes: https://github.com/ceph/ceph-salt/issues/208
Signed-off-by: Nathan Cutler <ncutler@suse.com>