ceph-salt-formula: use pkg.installed in apply only
It's not a good idea to put the pkg.installed state in sls files that will be
run often. This is because pkg.installed triggers a "zypper refresh" every time
it is run, and "zypper refresh" (a) takes a long time to run and (b) is prone to
failure.

Since we know that "ceph-salt apply" must completely successfully for ceph-salt
to be usable, we can move all pkg.installed invocations there. Elsewhere,
a simple file.exists check is enough to assert that the relevant RPM package is
still installed on the minion.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1177171
Signed-off-by: Nathan Cutler <ncutler@suse.com>