ceph-salt-formula: give explicit IP addr when adding host
The upstream cephadm developers recently added a commit
- 0facfac91fd8f71e5a8b869d818e7c2b07b93516 - which was backported to pacific
and was found to be causing our "ceph orch host add" command to fail because we
weren't specifying the IP address explicitly.

While the developers are now discussing whether the command really should fail
in this way, it seems prudent to give our IP address here explicitly, since
mgr/cephadm is going to try to resolve the hostname, anyway.

Fixes: https://github.com/ceph/ceph-salt/issues/463

Signed-off-by: Nathan Cutler <ncutler@suse.com>