tests: add source dir as real directory to fake fs
We use Cryptodome for generating SSH keys, and that in turn tries to
dynamically load a shared library to do the work.  Recently (2021-09-19)
the load_pycryptodome_raw_lib() function was changed to use os.path.isfile()
to check if the shared library exists before loading it:

https://github.com/Legrandin/pycryptodome/commit/47bca2b38665696d932df487f49a17b240f9c886

The os.path.isfile() check fails when used inside a pytest environment
with a pyfakefs present (which of course we have in our unit tests).
We can work around this by calling add_real_directory() to add the real
source directory to the fakefs, so the os.path.isfile() test will pass.

Signed-off-by: Michael Fritch <mfritch@suse.com>
Signed-off-by: Tim Serong <tserong@suse.com>