Move ceph-salt-registry-json creation to container.sls
This rejigs things slightly so that /tmp/ceph-salt-registry-json
is created prior to the bootsrap code, because it's needed
not just during bootstrap, but also separately in ceph-admin.sls
which can be run outside the context of a bootstrap, e.g.
during an upgrade from a previous (non-orchestrator, package
based) deployment, or when re-running `ceph-salt apply` at any
later time.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1187015
Signed-off-by: Tim Serong <tserong@suse.com>