Add OS and Ceph version info to `ceph-salt status` output
This allows the admin to easily see that the same base OS
version and Ceph packages are installed across the cluster.
Initially, before deployment, you'll see something like this:

  # ceph-salt status
  Cluster:   1 minions, 0 hosts managed by cephadm
  OS:        SUSE Linux Enterprise Server 15 SP2
  Ceph RPMs: Not installed
  Config:    OK

Later, when deployed:

  # ceph-salt status
  Cluster:   4 minions, 4 hosts managed by cephadm
  OS:        SUSE Linux Enterprise Server 15 SP2
  Ceph RPMs: ceph version 15.2.15-83-gf72054fa653
  Config:    OK

If you somehow manage to end up with mixed versions installed
while upgrading the base OS, you'll see something like this:

  master:~ # ceph-salt status
  Cluster:   4 minions, 4 hosts managed by cephadm
  OS:        Multiple versions running:
             - SUSE Linux Enterprise Server 15 SP3:
               - master.ses7.test
               - node1.ses7.test
             - SUSE Linux Enterprise Server 15 SP2:
               - node2.ses7.test
               - node3.ses7.test
  Ceph RPMs: Multiple versions installed:
             - ceph version 16.2.7-640-gceb23c7491b:
               - master.ses7.test
               - node1.ses7.test
             - ceph version 15.2.15-83-gf72054fa653:
               - node2.ses7.test
               - node3.ses7.test
  Config:    OK

In the case where there are multiple versions, the strings
"Multiple versions running" and "Multiple versions installed"
are rendered in orange to indicate that this is a slightly
strange but non-fatal situation.

Signed-off-by: Tim Serong <tserong@suse.com>